# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Search Ranking Platform
  version: 2.0.0
  description: Catalog, bounded search, attributed interactions, immutable data/model artifacts, deployment
    and experiments. Local service; administrative APIs require X-Admin-Token. Public routes do not verify
    end-user identity.
servers:
- url: http://127.0.0.1:8080
paths:
...
  /api/v1/search:
    post:
      operationId: search
      summary: Retrieve and rank bounded candidates
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      description: Creates a new request identity each time; does not record an exposure. A stopped experiment
        explicitly falls back to baseline. Score is a probability only when policy is model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
  /api/v1/events/impressions:
    post:
      operationId: recordImpression
      summary: Record an observed returned item
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      description: One exposure per request/ad. Identical ID and payload retry is 200; conflicting identity/content
        is 409. New exposure must arrive within one hour and occur in the hour after search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Impression'
  /api/v1/events/clicks:
    post:
      operationId: recordClick
      summary: Record a click attributed to an impression
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      description: Click must occur within the 24-hour window and arrive within one hour. Missing impression
        is 409. Several clicks do not multiply CTR labels.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Click'
  /api/v1/admin/features/snapshots:
    post:
      operationId: publishFeatures
      summary: Atomically publish immutable features
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      description: Unique ad IDs, known catalog references, a past cutoff and a nonempty window of at most
        90 days are required. Metadata and rows commit together.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeaturePublication'
  /api/v1/admin/features/latest:
    get:
      operationId: latestFeatures
      summary: Read latest available compatible snapshot
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureSnapshot'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
  /api/v1/admin/models:
    post:
      operationId: registerModel
      summary: Register an immutable model artifact
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRegistration'
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRegistration'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      description: All fields are mandatory. Feature order and finite parameters are checked. Evaluation cutoffs
        are strictly ordered and cannot end in the future. Does not promote the model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickModel'
    get:
      operationId: listModels
      summary: List up to 100 model versions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$
                maxItems: 100
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
  /api/v1/admin/models/{version}:
    get:
      operationId: getModel
      summary: Read a model artifact
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClickModel'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      parameters:
      - name: version
        in: path
        required: true
        schema:
          type: string
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$
  /api/v1/admin/deployment:
    get:
      operationId: getDeployment
      summary: Read current model selection and revision
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
    put:
      operationId: updateDeployment
      summary: Promote or roll back with optimistic concurrency
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      description: revision is the expected current revision. Success increments it. Null model requires zero
        rollout and selects the default baseline. Pinned experiments are independent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deployment'
  /api/v1/admin/deployment/history:
    get:
      operationId: deploymentHistory
      summary: Read up to 100 deployment revisions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                maxItems: 100
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
  /api/v1/admin/experiments:
    post:
      operationId: createExperiment
      summary: Create a frozen experiment definition
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentRegistration'
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentRegistration'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Experiment'
  /api/v1/admin/experiments/{id}/stop:
    post:
      operationId: stopExperiment
      summary: Stop new experiment traffic idempotently
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentStop'
        '400':
          description: Invalid request or contract
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Administrative token missing, invalid, or disabled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Referenced resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Identity, payload, or deployment revision conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Quality gate or supported size bound failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Database operation could not be confirmed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - AdminToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$
...
components:
  schemas:
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          items:
            type: object
            required:
            - field
            - message
            properties:
              field:
                type: string
              message:
                type: string
...
    SearchResponse:
      type: object
      additionalProperties: false
      required:
      - requestId
      - createdAt
      - policy
      - modelVersion
      - snapshotId
      - experimentId
      - arm
      - fallbackReason
      - items
      properties:
        requestId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        policy:
          type: string
          enum:
          - baseline
          - model
        modelVersion:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$
          - type: 'null'
        snapshotId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        experimentId:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$
          - type: 'null'
        arm:
          anyOf:
          - type: string
            enum:
            - control
            - treatment
          - type: 'null'
        fallbackReason:
          anyOf:
          - type: string
          - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/RankedItem'
          maxItems: 50
...
    Receipt:
      type: object
      additionalProperties: false
      required:
      - id
      - created
      properties:
        id:
          type: string
          format: uuid
        created:
          type: boolean
...
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Configured token of at least 32 characters. Administrative routes are disabled otherwise.